backend_python/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
python-dotenv==1.0.0
requests==2.31.0
psycopg2-binary==2.9.9
pgvector==0.2.4
langchain==0.1.0
langchain-community==0.0.10
huggingface-hub==0.19.4
sentence-transformers==2.2.2
torch==2.2.0
transformers==4.35.2
werkzeug==3.0.1
gunicorn==21.2.0
numpy>=1.20.0
setuptools
PyJWT==2.8.0
openai 
langchain-openai